Adds a solution hw04

+++ model/contacts.js
@@ -1,11 +1,34 @@
-const Contacts = require("./schemas/contact-schema");
-const listContacts = async () => {
-  const results = await Contacts.find({});
-  return results;
+const Contacts = require("./schemas/contact");
+
+const listContacts = async (
+  userId,
+  { sortBy, sortByDesc, filter, limit = "20", page = "1" }
+) => {
+  const results = await Contacts.paginate(
+    { owner: userId },
+    {
+      limit,
+      page,
+      sort: {
+        ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+        ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}),
+      },
+      select: filter ? filter.split("|").join(" ") : "",
+      populate: {
+        path: "owner",
+        select: "email subscription",
+      },
+    }
+  );
+  const { docs: contacts, totalDocs: total } = results;
+  return { total: total.toString(), limit, page, contacts };
 };
 
-const getContactById = async (id) => {
-  const result = await Contacts.findOne({ _id: id });
+const getContactById = async (id, userId) => {
+  const result = await Contacts.findOne({ _id: id, owner: userId }).populate({
+    path: "owner",
+    select: "email subscription",
+  });
   return result;
 };
 
@@ -14,19 +37,19 @@ const addContact = async (body) => {
   return result;
 };
 
-const updateContact = async (id, body) => {
+const updateContact = async (id, body, userId) => {
   const result = await Contacts.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true }
   );
   return result;
 };
 
-const removeContact = async (id) => {
-  const result = await Contacts.findByIdAndRemove({ _id: id });
+const removeContact = async (id, userId) => {
+  const result = await Contacts.findByIdAndRemove({ _id: id, owner: userId });
   return result;
-}
+};
 
 module.exports = {
   listContacts,
@@ -34,4 +57,4 @@ module.exports = {
   removeContact,
   addContact,
   updateContact,
-}
+};
